+++ webapp/image_processing/util.js
@@ -127,22 +127,6 @@ export function openPolygon(poly) {
     return poly;
 }
 
-// Convert polygons to martinez format
-// Input: [{ x: Number, y: Number }, ...]
-// Output: [ [ [x, y], ... ] ]
-export function toMartinez(poly) {
-    return [poly.map(pt => [pt.x, pt.y])];
-}
-
-// Input: [ [ [x, y], ... ] ]
-// Output: [{ x: Number, y: Number }, ...]
-export function fromMartinez(mtz) {
-    // mtz is array of rings: [ [ [x, y], ... ] ]
-    console.assert(Array.isArray(mtz));
-    console.assert(mtz.length == 1); // Only one ring
-    return mtz.map(ring => ring.map(([x, y]) => ({ x, y })));
-}
-
 // Draws polygons with bright colors, or a single override color if provided
 export function drawPolygonsWithColors(polygons, overrideColor = null) {
     const brightColors = [
