+++ webapp/pages/scene_image_claycode.html
@@ -9,7 +9,7 @@
   <script type="module" src="/scenes/scene_image_claycode.js"></script>
   <link rel="stylesheet" href="/image_processing/coloris.min.css" />
   <script src="/image_processing/coloris.min.js"></script>
-  <script src="../common/utils/martinez.min.js"></script>
+  <script src="../geometry/martinez.min.js"></script>
 </head>
 
 <body style="position: relative;">
